simplify the code back down to a trival command set before building it back up with an improved schema
